requirements.txt
# Core dependencies
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.30.0
numpy>=1.24.0
scipy>=1.10.0

# Computer Vision
opencv-python>=4.7.0
Pillow>=9.5.0

# Training and Logging
tqdm>=4.65.0
tensorboard>=2.13.0
matplotlib>=3.7.0

# Hugging Face
huggingface-hub>=0.26.2

# Cosmos Tokenizer dependencies
loguru>=0.7.0
mediapy==1.1.6
einops==0.7.0
einx==0.1.3

# Development and debugging
ipdb>=0.13.13

# Data handling
pandas>=2.0.0
